requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
pydantic==2.9.2
pypdf==4.2.0
sentence-transformers==3.0.1
chromadb==0.5.5
faiss-cpu==1.8.0.post1
structlog==24.1.0
tqdm==4.66.5
uvloop==0.19.0; platform_system != 'Windows'
uvicorn==0.30.6
orjson==3.10.7
numpy==1.26.4
